requirements.txt
vietocr
PyYAML
loguru
pillow
selenium
webdriver-manager
numpy==1.23.1
streamlit
Flask
flask_cors